imporve almost equal algorithm
We cannot say two numbers are almost equal without considering their application. 0.002 milimeter and 0.003 milimeter are almost equal if they are used in a product of the size of a house but we cannot say the same thing about them if they are used in a watch or micro robot. so physics matter and as we are focused on manufacturing industry where the most accurate cutters (wire cutters) have 1E-5 accuracy in meters. it is safe to assume if we round the value of the differences to 5 digits we would get a decent estimate of equality.