Dev (#7)
* pack interop project

* pack extensions project

* generated the nuget packages

* pack to nuget

* remove old nuget

* creat new nuget for extensions

* add extension nuget

* creat new nuget for addins

* creat new nuget for addins.fluent

* Add ref to new nugets

in Qrify  add reference to new nugets generated

* creat new nuget for addins.fluent

* make addin icons daily

* make addin icons daily

* remoe empty field

* add references

* fixed addin icons problem

* rebuilt extensions for new nugets

* imporve almost equal algorithm

* imporve almost equal algorithm

We cannot say two numbers are almost equal without considering their application. 0.002 milimeter and 0.003 milimeter are almost equal if they are used in a product of the size of a house but we cannot say the same thing about them if they are used in a watch or micro robot. so physics matter and as we are focused on manufacturing industry where the most accurate cutters (wire cutters) have 1E-5 accuracy in meters. it is safe to assume if we round the value of the differences to 5 digits we would get a decent estimate of equality.

* move Units.cs to own file

* fix typeos

* fix typos

* improve comment quality

* implimented logging

* add property to control when nugets are created

* remove qrifyplus installer

* Create sample wix installer

* revision is 2018.1

* revision is 2018.1

* Add evensource title to attribute

* change addin icon location

addin icon locaiton is best placed in the output folder so Wix and other installers can locate and then pulish it to the client machines. the default one that actually works is a bitmap 16x16

* build so the installers will be created

* sandbox log on commands

* build command

* remove the HKCU items

heat harvest the HKCU with the same components that contain HKLM. this causes error on wix compiler. we remove it from the generated file and manually add it to the component.wxs

* this removes the pdb and xml files from heat harvest output

* wix installer for Qrify

* must be marked as export to com

* update attribute to show the eventsource

* version is 2018.1

* version 2018.1