testCategory not category and save packages
there was a problem with test category, MSTEST (which i like more than xUnit) uses the TestCategory attribute not Category. also apparently you have to save the nuget files as artifacts if you want ot access them in the same job! why you may ask, I say because github is a tool to collect all the code in the world and provide minimum services.